refactor: address second review round
- Extract matchesStatusFilter to own .ts file (avoids React/Chakra import in tests)
- Fix describe("aggregateJobNameCounters") → describe("aggregateJobNameCounters()")
- Rename retryBlocked return { retried, unblocked } → { wasBlocked } (clearer semantics)
- Extract computeJobNameThroughput private method from computeRates (SRP)
